optimize the sequelizer to be a regular ole function
